requirements.txt
pandas
numpy
pyserial
